ENH: created `Converter` class
