+++ mosfit/strings.json
@@ -13,7 +13,7 @@
     "converged": "Converged!",
     "converting_to_json": "Converting input file `{}` to JSON format...",
     "copying": "Copying MOSFiT folder hierarchy to current working directory (disable with `--no-copy-at-launch`).",
-    "counts_or_mags": "\nNo header found in the specified datafile. Are the points presented in the datafile represented as magnitudes or as a count rate (or flux)?",
+    "counts_or_mags": "\nAmbigious header found in the specified datafile. Are the points presented in the datafile represented as magnitudes or as a count rate (or flux)?",
     "ctrl_c": "\b\bCtrl + C pressed, halting...",
     "dl_svo": "Downloading bandpass `{}` from SVO...",
     "dling_aliases": "Event `{}` interpreted as a transient name, downloading lists of transient aliases...",
@@ -58,7 +58,7 @@
     "no_processes": "MPI was unable to find any processes. This suggests that `mpi4py` was installed while another MPI library other than the one presently installed was active. Try re-installing either openmpi or mpich. Continuing in serial mode.",
     "no_ul_token": "No upload token found! Please visit {} to obtain an upload token for MOSFiT.",
     "omit_offline": "Try omitting the `--offline` flag.",
-    "one_per_line": "Does each row contain a single `{}`, or multiple `{}s` (e.g. a single row reports magnitudes for different `{}s` in different columns)?",
+    "one_per_line": "\nDoes each row contain a single `{}`, or multiple `{}s` (e.g. a single row reports magnitudes for different `{}s` in different columns)?",
     "paste_token": "Please paste your Dropbox token, then hit enter:",
     "pef_ext_search": "Could not find event, performing extended name search...",
     "progress": "Progress",
